requirements.txt
absl-py==0.7.1
alabaster==0.7.12
allennlp==0.8.3
appnope==0.1.0
asn1crypto==0.24.0
astor==0.8.0
atari-py==0.2.6
atomicwrites==1.3.0
attrs==19.1.0
aws-xray-sdk==0.95
awscli==1.16.141
Babel==2.6.0
backcall==0.1.0
benepar==0.1.2
blis==0.2.4
boto==2.49.0
boto3==1.9.98
botocore==1.12.131
box2d-py==2.3.8
bz2file==0.98
cachetools==3.1.1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.5.2
colorama==0.3.9
conllu==0.11
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
Cython==0.29.12
cytoolz==0.10.0
# de-core-news-sm==2.1.0
decorator==4.3.2
dgl==0.1.3
docker==3.7.2
docker-pycreds==0.4.0
docutils==0.14
ecdsa==0.13.3
editdistance==0.5.3
elkai==0.0.7
# en-core-web-sm==2.1.0
flaky==3.5.3
Flask==1.0.2
Flask-Cors==3.0.7
ftfy==5.5.1
future==0.17.1
gast==0.2.2
gensim==3.7.1
gevent==1.4.0
glfw==1.8.3
google-pasta==0.1.7
graphviz==0.10.1
greenlet==0.4.15
grpcio==1.22.0
gym==0.14.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
inflect==0.2.5
ipython==7.7.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.15.1
jellyfish==0.6.1
Jinja2==2.10.1
jmespath==0.9.3
joblib==0.13.2
jsondiff==1.1.1
jsonnet==0.12.1
jsonpickle==1.1
jsonschema==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
lockfile==0.12.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mock==2.0.0
more-itertools==7.0.0
moto==1.3.7
mpi4py==3.0.2
mpmath==1.1.0
msgpack==0.5.6
mujoco-py==2.0.2.5
murmurhash==1.0.2
networkx==2.3
nltk==3.4.4
nose==1.3.7
numpy==1.17.2
numpydoc==0.8.0
opencv-python==4.1.0.25
overrides==1.9
packaging==19.0
pandas==0.24.1
parsimonious==0.8.1
parso==0.5.1
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
plac==0.9.6
pluggy==0.9.0
plyfile==0.7
powerlaw==1.4.6
preshed==2.0.1
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.6.3
ptyprocess==0.6.0
PuLP==1.6.10
py==1.8.0
pyahocorasick==1.4.0
pyaml==19.4.1
pyasn1==0.4.5
pycairo==1.18.0
pycorenlp==0.3.0
pycparser==2.19
pycryptodome==3.8.1
pydot==1.4.1
pyemd==0.5.1
pygame==1.9.6
pyglet==1.3.2
Pygments==2.3.1
# PyGObject==3.30.4
# pyobjc==6.1
# pyobjc-core==6.1
# pyobjc-framework-Accounts==6.1
# pyobjc-framework-AddressBook==6.1
# pyobjc-framework-AppleScriptKit==6.1
# pyobjc-framework-AppleScriptObjC==6.1
# pyobjc-framework-ApplicationServices==6.1
# pyobjc-framework-Automator==6.1
# pyobjc-framework-AVFoundation==6.1
# pyobjc-framework-AVKit==6.1
# pyobjc-framework-CalendarStore==6.1
# pyobjc-framework-CFNetwork==6.1
# pyobjc-framework-CloudKit==6.1
# pyobjc-framework-Cocoa==6.1
# pyobjc-framework-Collaboration==6.1
# pyobjc-framework-ColorSync==6.1
# pyobjc-framework-Contacts==6.1
# pyobjc-framework-ContactsUI==6.1
# pyobjc-framework-CoreAudio==6.1
# pyobjc-framework-CoreAudioKit==6.1
# pyobjc-framework-CoreBluetooth==6.1
# pyobjc-framework-CoreData==6.1
# pyobjc-framework-CoreLocation==6.1
# pyobjc-framework-CoreMedia==6.1
# pyobjc-framework-CoreMediaIO==6.1
# pyobjc-framework-CoreML==6.1
# pyobjc-framework-CoreServices==6.1
# pyobjc-framework-CoreSpotlight==6.1
# pyobjc-framework-CoreText==6.1
# pyobjc-framework-CoreWLAN==6.1
# pyobjc-framework-CryptoTokenKit==6.1
# pyobjc-framework-DictionaryServices==6.1
# pyobjc-framework-DiscRecording==6.1
# pyobjc-framework-DiscRecordingUI==6.1
# pyobjc-framework-DiskArbitration==6.1
# pyobjc-framework-DVDPlayback==6.1
# pyobjc-framework-EventKit==6.1
# pyobjc-framework-ExceptionHandling==6.1
# pyobjc-framework-ExternalAccessory==6.1
# pyobjc-framework-FinderSync==6.1
# pyobjc-framework-FSEvents==6.1
# pyobjc-framework-GameCenter==6.1
# pyobjc-framework-GameController==6.1
# pyobjc-framework-GameKit==6.1
# pyobjc-framework-GameplayKit==6.1
# pyobjc-framework-ImageCaptureCore==6.1
# pyobjc-framework-IMServicePlugIn==6.1
# pyobjc-framework-InputMethodKit==6.1
# pyobjc-framework-InstallerPlugins==6.1
# pyobjc-framework-InstantMessage==6.1
# pyobjc-framework-Intents==6.1
# pyobjc-framework-IOSurface==6.1
# pyobjc-framework-iTunesLibrary==6.1
# pyobjc-framework-LatentSemanticMapping==6.1
# pyobjc-framework-LaunchServices==6.1
# pyobjc-framework-libdispatch==6.1
# pyobjc-framework-LocalAuthentication==6.1
# pyobjc-framework-MapKit==6.1
# pyobjc-framework-MediaAccessibility==6.1
# pyobjc-framework-MediaLibrary==6.1
# pyobjc-framework-MediaPlayer==6.1
# pyobjc-framework-MediaToolbox==6.1
# pyobjc-framework-MetalKit==6.1
# pyobjc-framework-ModelIO==6.1
# pyobjc-framework-MultipeerConnectivity==6.1
# pyobjc-framework-NetFS==6.1
# pyobjc-framework-NetworkExtension==6.1
# pyobjc-framework-NotificationCenter==6.1
# pyobjc-framework-OpenDirectory==6.1
# pyobjc-framework-OSAKit==6.1
# pyobjc-framework-Photos==6.1
# pyobjc-framework-PhotosUI==6.1
# pyobjc-framework-PreferencePanes==6.1
# pyobjc-framework-PubSub==6.1
# pyobjc-framework-QTKit==6.1
# pyobjc-framework-Quartz==6.1
# pyobjc-framework-SafariServices==6.1
# pyobjc-framework-SceneKit==6.1
# pyobjc-framework-ScreenSaver==6.1
# pyobjc-framework-ScriptingBridge==6.1
# pyobjc-framework-SearchKit==6.1
# pyobjc-framework-Security==6.1
# pyobjc-framework-SecurityFoundation==6.1
# pyobjc-framework-SecurityInterface==6.1
# pyobjc-framework-ServiceManagement==6.1
# pyobjc-framework-Social==6.1
# pyobjc-framework-SpriteKit==6.1
# pyobjc-framework-StoreKit==6.1
# pyobjc-framework-SyncServices==6.1
# pyobjc-framework-SystemConfiguration==6.1
# pyobjc-framework-VideoToolbox==6.1
# pyobjc-framework-Vision==6.1
# pyobjc-framework-WebKit==6.1
pyparsing==2.3.1
Pyphen==0.9.5
pytest==4.4.1
python-dateutil==2.8.0
python-jose==2.0.2
python-Levenshtein==0.12.0
pytorch-ignite==0.2.0
pytorch-pretrained-bert==0.6.1
pytorch-transformers==1.1.0
pyttsx3==2.71
pytz==2018.9
PyYAML==3.13
pyzmq==18.0.0
rdflib==4.2.2
regex==2019.4.14
repoze.lru==0.7
requests==2.21.0
responses==0.10.6
rsa==3.4.2
s3transfer==0.2.0
sacremoses==0.0.34
scikit-learn==0.20.3
scipy==1.3.1
seaborn==0.9.0
sentencepiece==0.1.82
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.8.0
snowballstemmer==1.2.1
spacy==2.1.3
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
# -e git+https://github.com/openai/spinningup.git@2e0eff9bd019c317af908b72c056a33f14626602#egg=spinup
sqlparse==0.3.0
srsly==0.0.5
stanfordnlp==0.2.0
tensorboard==1.14.0
tensorboardX==1.8
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
textacy==0.8.0
textstat==0.5.6
thinc==7.0.4
toolz==0.10.0
# torch==1.2.0
# torch-cluster==1.4.3
# torch-geometric==1.3.0
# torch-scatter==1.3.1
# torch-sparse==0.4.0
# torch-spline-conv==1.1.0
# torchfile==0.1.0
# torchvision==0.2.1
# torchviz==0.0.1
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
transformers==2.0.0
tsp==0.0.6
ujson==1.35
Unidecode==1.0.23
Unipath==1.1
urllib3==1.24.2
visdom==0.1.8.8
wasabi==0.2.1
wcwidth==0.1.7
websocket-client==0.54.0
Werkzeug==0.15.2
word2number==1.1
wrapt==1.11.1
xmltodict==0.12.0
you-get==0.4.1270
